fake_news_backend/requirements.txt
Flask
opencv-python
pillow
requests
